+++ lib/presentation/ui/page/user/navigation_page/mypage/account/component/user_profile_provider.dart
@@ -7,6 +7,3 @@ final userNameProvider = StateProvider.autoDispose((ref) {
   return TextEditingController(text: '');
 });
 
-final phoneNumberProvider = StateProvider.autoDispose((ref) {
-  return TextEditingController(text: '');
-});
